Change the argument used for device name and keyid
The order of the dm-android-verity params is change.
The code uses argument 0 as keyid and argument 1 as
device name. This cause the device boot failure.

Actually as per the below dm command, we changed to use
arg 0 as device name and arg 1 as keyid.

The commandline argumet for dm as below
dm="system none ro,0 1 android-verity
/dev/mmcblk0p62 Android:#7e4333f9bba00adfe0ede979e28ed1920492b40f"

Change-Id: I1cce12825966865b0dab4bfec9c79858fbe6ad41
Signed-off-by: Kanagarathina Kalathi <kalathik@motorola.com>
Reviewed-on: https://gerrit.mot.com/1391156
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Shi-Yong Li <a22381@motorola.com>
Reviewed-by: Ling Jin <lingjin@motorola.com>
Submit-Approved: Jira Key